+++ settings.php
@@ -54,7 +54,7 @@ if ($hassiteconfig) {
             'local_lumination/baseurl',
             get_string('setting_baseurl', 'local_lumination'),
             get_string('setting_baseurl_desc', 'local_lumination'),
-            'https://stage.ai-tutor.ai/api/v1',
+            'https://ai-tutor.ai/api/v1',
             PARAM_URL
         )
     );